app/utils/highlight_utils.py
from PIL import Image, ImageDraw
import fitz
import re
from typing import List, Dict, Any
from pathlib import Path

class HighlightEngine:
    def __init__(self, output_dir: str = "output_images"):
        self.output_dir = Path(output_dir)
        self.output_dir.mkdir(parents=True, exist_ok=True)

    def process_document(self, pdf_path: Path, answers: List[str]) -> List[Path]:
        doc = fitz.open(pdf_path)
        output_paths = []
        
        for page_num in range(len(doc)):
            page = doc.load_page(page_num)
            if highlighted := self._process_page(page, answers):
                output_paths.append(highlighted)
        
        doc.close()
        return output_paths

    def _process_page(self, page, answers: List[str]) -> Optional[Path]:
        # Highlighting implementation
        pass

    @staticmethod
    def extract_organizations(answer: str) -> List[str]:
        # Organization extraction logic
        pass

def create_highlight(text: str, start: int, end: int, label: str) -> Dict[str, Any]:
    """
    Create a highlight object for text.
    
    Args:
        text: The text to highlight
        start: Start position of highlight
        end: End position of highlight
        label: Label for the highlight
        
    Returns:
        Dict containing highlight information
    """
    return {
        "text": text,
        "start": start,
        "end": end,
        "label": label
    }

def merge_highlights(highlights: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
    """
    Merge overlapping highlights.
    
    Args:
        highlights: List of highlight objects
        
    Returns:
        List of merged highlight objects
    """
    if not highlights:
        return []
    
    # Sort highlights by start position
    sorted_highlights = sorted(highlights, key=lambda x: x["start"])
    merged = [sorted_highlights[0]]
    
    for highlight in sorted_highlights[1:]:
        prev = merged[-1]
        if highlight["start"] <= prev["end"]:
            # Merge overlapping highlights
            prev["end"] = max(prev["end"], highlight["end"])
            prev["text"] = prev["text"] + " " + highlight["text"]
        else:
            merged.append(highlight)
    
    return merged
