requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
llama-index-core>=0.10.20
llama-index-llms-huggingface>=0.5.0
sentence-transformers>=2.6.1
torch>=2.2.1
PyMuPDF>=1.24.0
pillow>=10.2.0
python-dotenv>=1.0.0
redis>=5.0.0
